+++ package.json
@@ -223,6 +223,7 @@
   "dependencies": {
     "@department-of-veterans-affairs/formation": "^1.11.0",
     "@department-of-veterans-affairs/react-jsonschema-form": "^1.0.0",
+    "@octokit/rest": "^15.17.0",
     "babel-plugin-dynamic-import-node": "^1.2.0",
     "blob-polyfill": "^2.0.20171115",
     "camelcase-keys-recursive": "^0.8.2",
